requirements.txt
aiohttp
async-timeout
attrs
beautifulsoup4
bs4
certifi
chardet
click
Flask
html-telegraph-poster
idna
itsdangerous
Jinja2
lxml
MarkupSafe
mega.py
multidict
pycryptodome
python-dotenv
pytz
requests
requests-toolbelt
six
soupsieve
telepot
tenacity
typing-extensions
urllib3
Werkzeug
yarl
cloudscraper==1.2.58